Add support for multiple possible types during partial evaluation
The exact type of a variable can often only be determined at runtime, and the PartialEvaluator can only
find out what the general supertype of all possible types is, e.g. for `Super s = flag ? new A() : new B()`.
For the analysis of all possible method calls however, this information is too imprecise. When using the new MultiTypedReferenceValueFactory, all possible types of a variable are tracked and available for further
analysis.