Add IncompleteClassHierarchyException
TypedReferenceValue previously threw a generic IllegalArgumentException
when a common superclass could not be found. This has been replaced with
a specific exception IncompleteHierarchyException, so that it can be
caught in a user's application to give further information.

ParallelAllClassVisitor no longers wraps the ExecutionException's cause
if it's already a RuntimeException to allow for catching exceptions like
IncompleteHierarchyException.