Add KotlinAnnotatable interface
Summary:
This diff introduces the `KotlinAnnotatable` interface. This will be used in the upcoming Kotlin annotation model visitors, to allow the context to be available in the `visit*` methods. For example:

```
    public void visitAnyAnnotation(Clazz               clazz,
                                   KotlinAnnotatable               annotatable,
                                   KotlinMetadataAnnotation annotation)
    {

    }
```

```
        public void visitIntValueAnnotationArgument(Clazz clazz, KotlinAnnotatable annotatable, KotlinAnnotation annotation, IntValue value)
        {

        }
```

Before this change we had `KotlinAnnotationVisitor.visitAnyAnnotation(Clazz, KotlinMetadataAnnotation)`; the context of what the annotation was applied to was not there, unless you implemented the specific methods.

You could (and still can) implement the individual interfaces like:

```
    default void visitTypeAnnotation(Clazz                    clazz,
                                     KotlinTypeMetadata       kotlinTypeMetadata,
                                     KotlinMetadataAnnotation annotation)
```

But this is not scalable when implement the Kotlin annotation model visitors for example, there would be a 3 * ~20 visit methods without a generic `KotlinAnnotatable`

```
        public void visitIntValueTypeAnnotationArgument(Clazz clazz, KotlinAnnotation annotation, IntValue value)
        public void visitIntValueTypeAliasAnnotationArgument(Clazz clazz, KotlinAnnotation annotation, IntValue value)
        public void visitIntValueTypeParameterAnnotationArgument(Clazz clazz, KotlinAnnotation annotation, IntValue value)
```