Fix CallResolver to not use unimplemented interface methods as viable targets
Summary: When searching for call targets in superinterfaces the call resolver previously didn't enforce that any potential target must be non-abstract. Also, if any other resolution path results in a call to an abstract method, this is now logged as an error to catch bugs earlier.

Test Plan: Existing tests

Reviewers: james.hamilton, carlo.alberto.pozzoli, dennis.titze

Reviewed By: dennis.titze

Differential Revision: http://phabricator.guardsquare.com/D10245