Extend callgraph to include possible subtypes of classes
The callgraph is completed with all possible types of methods. For each fied, parameter and retun value referenced class is searched in program and library classpools, then all subtypes of the class are collected and added as potential types, what results in callgraph that includes all possibilities. However if after collection of subtypes an java/lang/Object will be present in the set then no potential subtypes are added to avoid a huge blow-up of the callgraph.