Take inheritance into consideration for ParticularReferenceValue type
Summary: In ParticularReferenceValues's sanity checks we were just checking if the type corresponds exactly with the type of the value. This resulted in incorrect behavior for `checkcast` instructions when type corresponded to an ancestor class of value. This change takes this into consideration when `referencedClass` is known and we can check for inheritance.

Test Plan:
- all AppSweep tests pass
- amount of PartialEvaluator errors in AppSweep in my manual tests reduced considerably (will better evaluate this in the nightlies)
- amount of false positives for AppSweep's `LeakingKotlinAssertsEvaluator` reduced from 5 to 1 in my test, meaning we are able to correctly evaluate the value of more strings