Add differential representation of CPA maps
The differential representation stores maps as trees consisting of full maps in roots and actions (insertion/deletion) in other nodes. A map value is defined by applying a sequence of actions from the root to the current node to the root map. Actions are collapsed into root nodes on demand. Whenever a map API is used, the full map is reconstructed.

The current implementation does not provide an overall improvement due to computational complexity. It needs to be futher refined towards avoiding/optimizing the full map reconstruction. It can be achieved by caching and computing some of the map methods on the tree representation.