Fix DetailedArrayReferenceValue having side effects on previous values when updated
At each step of partial evaluation a new stack/variable array are created containing the same Value objects of the previous step. This is not a problem as long as the Value representation is immutable, but this is not the case for DetailedArrayReferenceValue for which the array is represented as an actual array since the values contained are calculated. By deep-copying the array before storing values and updating all its references on stack/variables the side affects are avoided