Add utility to do on-demand lightweight call resolution
When invoke instructions are encountered for which the `CallResolver` was unable to create a `Call` object (resource limits or other general resolution errors), you can now use `CallResolver#quickResolve` to do on-demand resolution of the selected instruction's call target. This method will not provide full target information in all cases, as e.g. type analysis for virtual calls is not executed.