Make `JvmTransferRelation` produce a successor if call edge is missing
Since we can't assume the call analysis to terminate we need to support the case in which interprocedural call edges are missing. In this case the transfer relation, instead of returning no successor, analyzes the calls intraprocedurally by performing a minimal call resolution. After the changes analyses with depth 0 (intraprocedural) can be performed without generating the inter-procedural CFA.