Refactor call class and fix invokedynamic resolution
Calls now incorporate some common logic that was previously part of utilities (e.g. checking if a call is static or what the total size of the arguments is), contain the full instruction they represent instead of only the opcode (for more post-resolution capabilities, e.g. in the future we might want to find out which lambda will be constructed through invokedynamic calls) and the resolution of invokedynamic calls now works as expected (we don't know the exact target, only the descriptor).

Closes: #63