Implement object tainting
Object tainting allows assigning a taint value to an object and all heap memory reachable from it. The object taint is stored in the empty field `""` abusing the absence of empty fields in Java. Whenever an object is tainted, all its fields are transitively tainted as values and objects. The object taint is used as the default value when new fields are discovered. Thus, whwnever a field is overwritten with another value, the object taint is ignored, which allows to reassign some fields of a tainted object to untainted values. Upon join, tainted objects transitively propagate their taint on mutually undiscovered fields. This models the presence of object taint in missing mappings which normally behave as the lattice bottom. One can select using object tainting by specifying `TAINT_TREE` as the heap model type.