Add limited hash maps for capping the tree heap model size
Limited hash maps decide on whether to skip a new mapping based on a tripredicate of the key, value, and the map itself. An abstract state backed by this maps is used on the top level of the tree heap model. Thus, CPA run can (dynamically) limit the size of the heap model.