Add field type to static field key
Summary:
The root cause of the following error is that fields can be overloaded with different types in Java class files:

```
 proguard.analysis.cpa.algorithms.CpaAlgorithm: CPA run stopped for the following error:  class proguard.evaluation.value.ParticularIntegerValue cannot be cast to class proguard.evaluation.value.IdentifiedReferenceValue (proguard.evaluation.value.ParticularIntegerValue and proguard.evaluation.value.IdentifiedReferenceValue are in unnamed module of loader 'app')
```

A class exists with many fields named `a` but with different types. At some point an integer is put in the integer field and a StringBuilder is put in the Object field. Previously, the key in the field map did not contain the type; this meant the field only existed once instead of multiple times with different types.