Fix bug that potentially causes methods clashing in hierarchy
Summary: This is to fix a bug that prevent methods from being renamed to a unique name when using the new renaming strategy (`RENAME_FIELD_IF_CLASH`). In the previous implementation of the `RENAME_FIELD_IF_CLASH` strategy, we only rename the field when it clash but the logic ignores method completely, leaving methods being not renamed. This causes funny issues like IllegalArgumentException and java.lang.LinkageError as the generalized method signature can clash with something in the hierarchy or even clash with something in the class scope.