PartialEvaluator support for String methods that return arrays
Improves the support for the behavior of methods that transform a `String` value into an array of values to now create an array reference (e.g. `DetailedArrayReferenceValue`), while before a `ParticularReferenceValue` was created.
A frequent example is a `String` being transformed using a method such as `toBytes()`/`toCharArray()`.
Tests were added and the `PartialEvaluatorHelper` was corrected to use the `DetailedArrayValueFactory`.

The `createArrayReferenceValue`* is being suggested to marked as deprecated because:
1. it is not being used in practice, the few implementations fallback on the version without the `elementValue`
2. it does not make much sense creating and array reference with a `Value` (not clear what the value can be)

*`ValueFactory.createArrayReferenceValue(String type, Clazz referencedClass, IntegerValue arrayLength, Value elementValue)`