Use actual type of ParticularReferenceValue in StringReflectionExecutor
Summary:
This makes `StringReflectionExecutor.getInstanceCopyIfMutable()` use the actual type of a reconstructed `ParticularReferenceValue` instead of the call target's type - the two types can be different, for example when code unsafely casts objects before calling methods.
This fixes a previous bug where a class cast exception was thrown as a result.