Support dynamic dispatch in Executors
Previously, we did not support dynamic dispatch in our Executors. For example, we were not able to analyze the following code with interprocedural value analysis:

```
Object key = "secretkey";
cipher.encrypt(key.toString())
```

The `key.toString()`would be resolved to `Object#toString()` and as we don't have any executors for that, the method would not be simulated. This diff adds support for resolving method call targets based on dynamically available information about the type of the instance objet. In the example above, we now correctly execute the `StringReflectionExecutor`.

Specifically, this diff introduces these changes:

1. When picking an executor for a method call, we try to find any compatible type our executors support by implementing similar lookup to what the JVM does. We try all types between the type provided by the method reference in the instruction (type lower bound) and the type of the instance object we actually observe (type upper bound).

2. The Executor API for expressing which methods are supported has been changed from imperative to declarative. Executors now have to provide a full list of methods that they support. Without this, the above checks can't be done reasonably fast.