Fix `ReflectionExecutor` not updating instance of `StringBuilder`s in fallback result
The executors should provide a full picture of the method execution. If something like:
```
StringBuilder a = new StringBuilder("Hello");
StringBuilder b = a.append(unknownString());
```
is analyzed, both `a` and `b` should have an unknown value. `ReflectionExecutor` was doing this correctly when the reflective execution happens, but not when there is some unknown parameter. In this case the return value (`b`) whould have been modified, but not the calling instance (`a`) and everything with its same reference.