Implement ClassLoader model with reflective methods
Follow up which implements a ClassLoader model and a number of its relevant API methods.
Moves those implementations out of the JavaReflectionApiExecutor as well.
The diff is a bit larger, since the model needs access to the class pools, in order to generate class models with references to the fetched classes.
This involves a small change to an API which is used in many of the tests.