Replace `JvmValueCpaRun` with `ValueAnalyzer`
This is the first step toward removing the entire `CpaRun` hierarchy, while making the utility classes to run the analysis more simple to read and use. Once `TaintAnalyzer` has also been added, the full hierarchy of `CpaRun` will be removed.

There are different ideas behind this refactor:
- since the intermediate hierarchy for `CpaRun` is not useful, since in practice it just make everything difficult to understand with an improvement in code reuse that is insignificant, this flattens all components of `JvmValueBamCpaRun` in a unique utility
- configuration options that are in practice not very useful have been removed. In case of more advanced use cases these options can either be reintroduced in the future, or the analysis can be run directly with the CPA classes without the need for a utility
- the old CpaRun code made not possible, or at best clunky and not clear, running different analyses while using the same cache to share the results. These changes allow explicitly to either build a new analyzer with a fresh cache, or to run the same analyzer several times with the same cache

The behavior that is not too clean at the moment for this code is how `ValueAnalysisResult` works, since having a result with side effects for following runs might not be ideal. It could be argued that, until providing a view on the results of a single execution of `analyze` is not possible, it would be better to have `analyze` return `void` while providing an explicit method to retrieve the (now more clearly global) result. While there is room for discussion the current solution should provide a better API for future improvements, while clearly documenting the behavior that might be unexpected.