Read libraryClassPool in ClassPools pair lazily
Summary:
When `val (programClassPool, libraryClassPool) = ClasspoolBuilder.fromSource(...)` is used, the destructuring happens via `component1()` and `component2()` methods. If a field is `_`, its `component` method is not called. By rewiring the `ClassPools` class, we can make sure the `libraryClassPool` can be fetched lazily, i.e. not fetched at all if `val (programClassPool, _)` is used.
Note that fetching the library class pool will still happen if the `initialize` parameter of `fromSource` is true, which is the default. However, it does open the door for single test runs that don't need initialization to boot 2 or 3 seconds faster.